refactor(stores): move kv and jobs subpackages (#1960 slice 3)
