+++ src/factory/bootstrap/standalone/hub_standalone.py
@@ -206,7 +206,7 @@ async def _bootstrap_hub_standalone(  # noqa: C901, PLR0915 — DEBT:migration-s
         # Provision active-jobs KV bucket before announcing readiness.
         # Workers / adapters consulting the registry rely on the bucket
         # existing before they receive the hub-ready signal. ADR-079 S3.
-        from factory.infrastructure.stores.active_jobs_kv import ensure_active_jobs_kv
+        from factory.infrastructure.stores.jobs.active_jobs_kv import ensure_active_jobs_kv
 
         try:
             await ensure_active_jobs_kv(_audio_js)
@@ -224,8 +224,8 @@ async def _bootstrap_hub_standalone(  # noqa: C901, PLR0915 — DEBT:migration-s
         _dlq_router = DlqRouter(nc, _audio_js)
         await _dlq_router.start()
 
-        from factory.infrastructure.stores.active_jobs_kv import KvActiveJobsStore
-        from factory.infrastructure.stores.active_jobs_refresher import (
+        from factory.infrastructure.stores.jobs.active_jobs_kv import KvActiveJobsStore
+        from factory.infrastructure.stores.jobs.active_jobs_refresher import (
             RegistryCoordinator,
         )
