+++ src/factory/infrastructure/stores/kv/__init__.py
@@ -1,0 +1,5 @@
+"""NATS KV-backed index stores."""
+
+from factory.infrastructure.stores.kv.message_index_kv import MessageIndexKvStore
+
+__all__ = ["MessageIndexKvStore"]
