+++ src/factory/infrastructure/stores/kv/message_index_kv.py
@@ -14,7 +14,7 @@ import nats.errors
 from nats.js.api import KeyValueConfig, StorageType
 from nats.js.errors import BadRequestError, KeyNotFoundError
 
-from factory.infrastructure.stores._kv_keys import kv_safe_part
+from factory.infrastructure.stores.kv._kv_keys import kv_safe_part
 
 if TYPE_CHECKING:
     from nats.js.client import JetStreamContext
